+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - add link to the wiki page for "first pane/window issue" to the README as well
   as other tweaks
 - save and restore grouped sessions (used with multi-monitor workflow)
+- save and restore active and alternate windows in grouped sessions
 
 ### v1.5.0, 2014-11-09
 - add support for restoring neovim sessions